+++ app/Domits/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "react": "18.2.0",
-    "react-native": "0.73.4"
+    "react-native": "^0.73.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
