+++ web/src/ui-components/index.js
@@ -4,5 +4,11 @@
  * Any changes to this file will be overwritten when running amplify pull. *
  **************************************************************************/
 
+export { default as AccommodationCreateForm } from "./AccommodationCreateForm";
+export { default as AccommodationUpdateForm } from "./AccommodationUpdateForm";
 export { default as Property1Opened } from "./Property1Opened";
+export { default as TodoCreateForm } from "./TodoCreateForm";
+export { default as TodoUpdateForm } from "./TodoUpdateForm";
+export { default as UserCreateForm } from "./UserCreateForm";
+export { default as UserUpdateForm } from "./UserUpdateForm";
 export { default as studioTheme } from "./studioTheme";
